Handle some palette textures on the GPU. vulkan: some linear tiling

+++ core/rend/gl4/abuffer.cpp
@@ -423,7 +423,7 @@ void DrawTranslucentModVols(int first, int count)
 
 	int mod_base = -1;
 
-	for (u32 cmv = 0; cmv < count; cmv++)
+	for (int cmv = 0; cmv < count; cmv++)
 	{
 		ModifierVolumeParam& param = params[cmv];
 
@@ -432,7 +432,7 @@ void DrawTranslucentModVols(int first, int count)
 
 		u32 mv_mode = param.isp.DepthMode;
 
-		verify(param.first >= 0 && param.first + param.count <= pvrrc.modtrig.used());
+		verify(param.first >= 0 && param.first + param.count <= (u32)pvrrc.modtrig.used());
 
 		if (mod_base == -1)
 			mod_base = param.first;
